move code to fermion_operator.py
